+++ GW2EIBuilders/JsonModels/JsonActors/JsonPlayer.cs
@@ -61,14 +61,14 @@ namespace GW2EIBuilders.JsonModels
         public List<int>[][] TargetDamage1S { get; internal set; }
         [JsonProperty]
         /// <summary>
-        /// Array of int representing 1S breakbar damage points \n
+        /// Array of double representing 1S breakbar damage points \n
         /// Length == # of targets and the length of each sub array is equal to # of phases
         /// </summary>
         /// <remarks>
         /// If the duration of the phase in seconds is non integer, the last point of this array will correspond to the last point  \n
         /// ex: duration === 15250ms, the array will have 17 elements [0, 1000,...,15000,15250]
         /// </remarks>
-        public List<int>[][] TargetBreakbarDamage1S { get; internal set; }
+        public List<double>[][] TargetBreakbarDamage1S { get; internal set; }
         [JsonProperty]
         /// <summary>
         /// Per Target Damage distribution array \n
@@ -208,7 +208,7 @@ namespace GW2EIBuilders.JsonModels
             {
                 NPC target = log.FightData.Logic.Targets[j];
                 var graph1SDamageList = new List<int>[phases.Count];
-                var graph1SBreakbarDamageList = new List<int>[phases.Count];
+                var graph1SBreakbarDamageList = new List<double>[phases.Count];
                 var targetDamageDistList = new List<JsonDamageDist>[phases.Count];
                 for (int i = 0; i < phases.Count; i++)
                 {
